pyproject.toml
[tool.poetry]
name = "commits-dataset"
version = "0.1.0"
description = ""
authors = ["Aleksandra Eliseeva"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8,<3.11"
tqdm = "~4.62.1"
hydra-core = "1.1.1"
jsonlines = "~3.0.0"
numpy = "~1.21.5"
pandas = "~1.5.1"
nltk = "3.6.6"

[tool.poetry.group.exact_hash.dependencies]
pyspark = "^3.3.2"

[tool.poetry.group.collection.dependencies]
pydriller = "1.15.5"

[tool.poetry.group.tokenizer.dependencies]
tokenizers = "~0.12.1"
transformers = "~4.21.1"

[tool.poetry.group.dev.dependencies]
black = "~22.6.0"
isort = "~5.10.1"
mypy = "~0.971"
pytest = "~7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
